refactor: 엔티티 파일 위치 변경
